Reorganize python module once more splitting of config
