+++ python/ddml/__init__.py
@@ -2,22 +2,22 @@ from typing import List
 import argparse
 import sys
 
-from .common import add_shower_model, user_physics, ModelPreset
-from . import presets
+from .setup_physics import add_shower_model, ddml_physics
+from .preset import ModelPreset
 
 
-PRESETS: dict = {}
+_PRESETS: dict = {}
 
 
 def _register_presets() -> None:
-    from . import presets as _presets_mod
+    from . import configs as _config_mod
 
-    for attr in dir(_presets_mod):
+    for attr in dir(_config_mod):
         if attr.startswith("_"):
             continue
-        val = getattr(_presets_mod, attr)
+        val = getattr(_config_mod, attr)
         if isinstance(val, ModelPreset):
-            PRESETS[attr] = val
+            _PRESETS[attr] = val
 
 
 _register_presets()
@@ -28,7 +28,7 @@ def get_presets_from_args() -> List[ModelPreset]:
     # ddsim). We do this to make it possible to dynamically get preests from CLI
     _cli = argparse.ArgumentParser(add_help=False)
     _cli.add_argument(
-        "--ml-preset",
+        "--ml-model",
         action="append",
         default=None,
         help="Preset dotted name (e.g. torch.CALOCLOUDS). Repeat to compose.",
@@ -39,16 +39,15 @@ def get_presets_from_args() -> List[ModelPreset]:
     sys.argv[:] = [sys.argv[0]] + remainder
 
     presets = []
-    for preset in args.ml_preset:
-        presets.append(PRESETS[preset])
+    for preset in args.ml_model:
+        presets.append(_PRESETS[preset])
     return presets
 
 
 __all__ = [
     "ModelPreset",
-    "PRESETS",
     "add_shower_model",
-    "user_physics",
+    "ddml_physics",
     "presets",
     "get_presets_from_args",
 ]
