+++ python/ddml/geometry.py
@@ -11,21 +11,7 @@ class DetectorGeometry:
     symmetry: Optional[int] = None  # The symmetry (if applicable) of the geometry
 
 
-_ILD_ECAL_BARREL = DetectorGeometry(
-    detector="EcalBarrel", region="EcalBarrelRegion", symmetry=8
-)
-_ILD_ECAL_ENDCAP = DetectorGeometry(detector="EcalEndcap", region="EcalEndcapRegion")
-_ILD_HCAL_BARREL = DetectorGeometry(
-    detector="HcalBarrel", region="HcalBarrelRegion", symmetry=8
-)
-_ILD_HCAL_ENDCAP = DetectorGeometry(detector="HcalEndcap", region="HcalEndcapRegion")
-
-
 @dataclass(frozen=True)
 class PluginGeometry:
     ecal: DetectorGeometry
     hcal: Optional[DetectorGeometry] = None
-
-
-ILD_BARREL = PluginGeometry(ecal=_ILD_ECAL_BARREL, hcal=_ILD_HCAL_BARREL)
-ILD_ENDCAP = PluginGeometry(ecal=_ILD_ECAL_ENDCAP, hcal=_ILD_HCAL_ENDCAP)
